Fix multi-threaded benchmarking
This fix enables fio threads to write a sum of 512MB of data in a 512MB
file. Previously, it was only possible for threads to write a sum of
512MB of data to a 512MB/num_threads file. This would falsely lead to
better results as the thread number went higher since the file would be
smaller and thus the write area smaller, leading to better locality of
writes.