Major code revision
This commit marks the attempt to unify all my benchmark scripts
(i.e. have the same starting point) in order to avoid writing the
same piece of code in two/three essentially same files.

Also, this commit offers a way to batch benchmarks via the batchos
script, which will be updated to reflect the above changes.